Rename SonarModelState to SonarState
